# perf: Use Span CopyTo instead of Array CopyTo for better performance